+++ substra/cli/printers.py
@@ -253,15 +253,24 @@ class AssetPrinter(BasePrinter):
 class JsonPrinter:
     @staticmethod
     def print(data, *args, **kwargs):
-        print(json.dumps(data.dict(), indent=2, default=str))
+        if isinstance(data, list):
+            for data_i in data:
+                print(json.dumps(data_i, indent=2, default=str))
+        else:
+            print(json.dumps(data.dict(), indent=2, default=str))
 
 
 class YamlPrinter:
     @staticmethod
     def print(data, *args, **kwargs):
         # We need the yaml format to display the same things than json format
-        json_format = json.dumps(data.dict(), indent=2, default=str)
-        print(yaml.dump(json.loads(json_format), default_flow_style=False))
+        if isinstance(data, list):
+            for data_i in data:
+                json_format = json.dumps(data_i, indent=2, default=str)
+                print(yaml.dump(json.loads(json_format), default_flow_style=False))
+        else:
+            json_format = json.dumps(data.dict(), indent=2, default=str)
+            print(yaml.dump(json.loads(json_format), default_flow_style=False))
 
 
 class AlgoPrinter(AssetPrinter):